Removed `#[derive(Copy)]` trait and `#[repr(packed)]`.
The packed representation requires the `Copy` trait. However, this should
not be necessary as the structs were already hand-optimized to compress the
struct fields as best as possible in there C representation (which should
not even matter now that they are never sent to the C-side library).

Removing the `Copy` trait prevents copying the event when pushing it onto
the `TRACES` static vector.