Major design change: `Register` trait
The `interpol-rs` crate now provides the public `Register` trait that allows library-defined MPI events to be registered in a static vector that lives for the entire program's execution.

As opposed to the previous version of `interpol-rs` where we used an enum and static dispatch to push events into the vector, the library now uses the concept of trait objects and performs dynamic dispatch (i.e. at runtime) to add traces of interposed MPI to the vector. Although this has performance consequences, it dramatically increases the flexibility of the code and makes it much more robust, easier to maintain and refactor for future additions. It also makes the library more consistent as each MPI events defined in it can implement a custom behavior when being registered in the static vector, while also abstracting the latter in a generic way, thus simplifying how we add support for other MPI routines we would like to trace.

The implementation of the `extern "C"` as well as the implementations of the `Register` trait on MPI events prevents the program from panicking when trying to allocate more memory. As this tools is designed to benchmark large, production-scale MPI applications, it is not desirable to stop the execution of the said program because of an error caused by the library. There are only three cases where the program may panic:
- If the program fails to take the lock on the mutex guarding the static vector of events (i.e. because a thread panicked/was poisoned while holding the lock);
- If the serialization of the vector of events fails;
- If writing the serialized contents to a output file fails.
These last two points should not be a problem for the user as its MPI program should have finished running by the point these errors could be raised.